Changed App.xaml.cs to extend Application, Added composition to CaliburnMicro to override the App.xaml.cs file, Changed postactions on app.xaml.cs file to be added by composition templates Changed path of tests to avoid errors caused by too long path
