+++ templates/_composition/CaliburnMicro/Feature.BackgroundTask.AddOnBackgroundNavigated/App_postaction.xaml.cs
