+++ templates/_composition/CaliburnMicro/Feature.SuspendAndResume.AddOnBackgroundEntered/.template.config/template.json
@@ -1,0 +1,28 @@
+﻿{
+  "author": "Microsoft",
+  "classifications": [
+    "Universal"
+  ],
+  "name": "CaliburnMicro.Feature.SuspendAndResume.AddOnBackgroundEntered", 
+  "tags": {
+    "language": "C#",
+    "type": "item",
+    "wts.type": "composition",
+    "wts.version": "1.0.0",
+    "wts.compositionFilter": "$framework == CaliburnMicro & identity == wts.Feat.SuspendAndResume"
+  },
+  "sourceName": "wts.ItemName",
+  "preferNameDirectory": true,
+  "PrimaryOutputs": [
+  ],
+  "symbols": {
+    "wts.rootNamespace": {
+      "type": "parameter",
+      "replaces": "Param_RootNamespace"
+    },
+    "wts.itemNamespace": {
+      "type": "parameter",
+      "replaces": "Param_ItemNamespace"
+    }
+  }
+}
