+++ templates/_composition/CaliburnMicro/Feature.SuspendAndResume.AddOnBackgroundEntered/App_postaction.xaml.cs
