+++ templates/_composition/CaliburnMicro/Project/App.xaml.cs
@@ -1,15 +1,16 @@
 ﻿using System;
 using Windows.ApplicationModel.Activation;
 using Windows.UI.Xaml;
+using System.Collections.Generic;
+using Caliburn.Micro;
+using Param_RootNamespace.Services;
 
-using wts.DefaultProject.Services;
-
-namespace wts.DefaultProject
+namespace Param_RootNamespace
 {
     /// <summary>
     /// Provides application-specific behavior to supplement the default Application class.
     /// </summary>
-    public sealed partial class App : Application
+    public sealed partial class App
     {
         private Lazy<ActivationService> _activationService;
 
@@ -26,6 +27,8 @@ namespace wts.DefaultProject
         public App()
         {
             InitializeComponent();
+
+            Initialize();
 
             // Deferred execution until used. Check https://msdn.microsoft.com/library/dd642331(v=vs.110).aspx for further info on Lazy<T> class.
             _activationService = new Lazy<ActivationService>(CreateActivationService);
@@ -51,6 +54,43 @@ namespace wts.DefaultProject
         protected override async void OnActivated(IActivatedEventArgs args)
         {
             await ActivationService.ActivateAsync(args);
+        }
+
+        private WinRTContainer _container;
+
+        /// <summary>
+        /// Configuration for Caliburn.Micro
+        /// </summary>
+        protected override void Configure()
+        {
+            // This configures the framework to map between MainViewModel and MainPage
+            // Normally it would map between MainPageViewModel and MainPage
+            var config = new TypeMappingConfiguration
+            {
+                IncludeViewSuffixInViewModelNames = false
+            };
+
+            ViewLocator.ConfigureTypeMappings(config);
+            ViewModelLocator.ConfigureTypeMappings(config);
+
+            _container = new WinRTContainer();
+            _container.RegisterWinRTServices();
+
+        }
+
+        protected override object GetInstance(Type service, string key)
+        {
+            return _container.GetInstance(service, key);
+        }
+
+        protected override IEnumerable<object> GetAllInstances(Type service)
+        {
+            return _container.GetAllInstances(service);
+        }
+
+        protected override void BuildUp(object instance)
+        {
+            _container.BuildUp(instance);
         }
     }
 }
