+++ templates/_composition/MVVMLight/Project.TabbedPivot/App_postaction.xaml.cs
@@ -1,6 +1,6 @@
 ﻿namespace Param_RootNamespace
 {
-    public sealed partial class App
+    public sealed partial class App : Application
     {
 //^^
 //{[{
