+++ templates/_composition/_shared/Feature.BackgroundTask.AddOnBackgroundNavigated/App_postaction.xaml.cs
@@ -1,0 +1,17 @@
+﻿//{**
+// This code block adds the OnBackgroundActivated event handler to the App class of your project.
+//**}
+namespace Param_RootNamespace
+{
+    public sealed partial class App : Application
+    {
+//^^
+//{[{
+
+        protected override async void OnBackgroundActivated(BackgroundActivatedEventArgs args)
+        {
+            await ActivationService.ActivateAsync(args);
+        }
+//}]}
+    }
+}
